requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
opencv-python
deepface
google-cloud-storage
numpy
google-auth
yq
jq
psycopg2-binary
functions-framework
python-dotenv
scikit-learn